Allow passing extra arguments along to the lean subprocess.
Does so by allowing writing e.g.:

    $ lean-language-server --stdio -- -M 4096 -T 10000

which will provide the extra argv down to Lean.

This is 'wrong' in the sense that it should really use the LSP-native
way to do so, which is responding to workspace/configuration requests,
but the current version of lean-language-server uses a quite old version
of vscode-languageserver, and I couldn't get it to work on a newer one
without significant changes. For another time...